Update logic for last day of month in datetimekeepr
This adds logic so that we never do an overflow for days in the
datetimekeeper. The new logic is such that if we do a change of month
(month delta) and it's on the last day of the month, if the new month would
have less days we just pick the last day of that month.

If the new month has more days we keep the current day of month