Add option for DuckDuckGo email address workflows
This adds an option to generate `duck` email addresses to be used for
keeping the main address private.

It also allows for formatting an existing email address to one that DDG
can handle. This allows the sender to initiate an email without having
to go through the process of replying to it first.

If neither option is wanted then it allows for just using the default
provided address for DDG.