Added CLI option to interact with mortgage server (#57)
This adds a new option to allow interacting with the upstream server that
handles tracking and management of the various mortage services. This
provides a CLI interface to the server so we dont need to keep hitting it
with curl and Web commands

This also updates the CI system to have more robust test cases and fixes a
bunch of cargo clippy errors that were showing up but not getting surfaced.